+++ docs/history.rst
@@ -1,6 +1,11 @@
 Release history
 ===============
 
+`Next Release`_
+---------------
+- Add :data:`~klempner.config.URL_SCHEME_MAP` so that URL schemes can be
+  easily configured.
+
 `0.0.1`_ (5 May 2019)
 ---------------------
 - Implement :ref:`simple-discovery-method`, :ref:`consul-discovery-method`,